Movie and watchlist repositories (#3)
* Domain class(es) and MovieRepository

- added Movie domain class
- added DataSource and Repository interface/abstract class in Persistence file
- implemented Repository pattern for Movies as MovieRepository
- KtxUtils to hold the new toImmutableList() extension function

* Finished MovieRepository

Added method for populating it which is called when it's first constructed.
Also, updated the Movie a little bit so it now properly overrides equals() and hashCode()
which will be necessary in the WatchlistRepository

* Add more javadoc to MovieRepository

* Wrapped up WatchlistRepository

Also had to extend the DataSource with removeItem, which subsequently affected
the other repo. A new extension function .toImmutableSet() has also been made.

* Fix drawables that were breaking the build

* Some code shuffling around and initial MovieRepo tests

+ extracted createLocalDate from MovieRepository to KtxUtils
+ fixed a bug in PreO version to append 0 when necessary to produce dates such as 2020-09-05
+ added drawable to Movie class so it can render properly in the recyclerview
+ initial MovieRepositoryTest that checks whether the repository is nicely prepopulated
  works fine, more tests to come

* Initial set of tests are in place

Even though the tests don't really test for anything as they're not using mocks
or anything fancy, they do have one sanity test so that more tests can be added
later.