Move MovieListFragment to it's own package
And also include a new MovieAdapter, which is a ListAdapter which is really just a fancier
RecyclerView adapter which lets us use the new "submitList()" method which diffs the new list
against the old one on a background thread and updates data without explicitly calling
notifyDataSetChanged(). Not much of a save on 6 movies but would be useful when the movie list
gets expanded.