The rest of the code
Launching MovieDetails by clicking on a movie from the MovieList's recycler view works great.
Adding/removing to watchlist repo also works great, and is reflected on the MovieList's items.

however, pressing back to go back to the MovieListFragment doesn't work.
neither is the menu completed.
The "on my watchlist" label in the recyclerview's items isn't really like it's in the picture
there is no sorting from the menu
there is no "back" carret in the toolbar once on the MovieDetails